refactor: move Copilot and its helper file to dedicated directory
