+++ lua/codecompanion/adapters/copilot/init.lua
@@ -1,4 +1,5 @@
 local config = require("codecompanion.config")
+local copilot_helper = require("codecompanion.adapters.copilot.helpers")
 local curl = require("plenary.curl")
 local log = require("codecompanion.utils.log")
 local openai = require("codecompanion.adapters.openai")
@@ -27,11 +28,6 @@ local utils = require("codecompanion.utils.adapters")
 ---@field vsc_electron_fetcher boolean
 ---@field xcode boolean
 ---@field xcode_chat boolean
-
-local _cached_adapter
-local _cache_expires
-local _cache_file = vim.fn.tempname()
-local _cached_models
 
 ---@alias CopilotOAuthToken string|nil
 local _oauth_token
@@ -154,206 +150,6 @@ local function get_and_authorize_token(self)
   return true
 end
 
----Reset the cached adapter
----@return nil
-local function reset()
-  _cached_adapter = nil
-end
-
----Get a list of available Copilot models
----@param self CodeCompanion.Adapter
----@param opts? table
----@return table
-local function get_models(self, opts)
-  if _cached_models and _cache_expires and _cache_expires > os.time() then
-    return _cached_models
-  end
-
-  if not _cached_adapter then
-    if not self then
-      return {}
-    end
-    _cached_adapter = self
-  end
-
-  get_and_authorize_token(self)
-  local url = _github_token.endpoints.api or "https://api.githubcopilot.com"
-  local headers = vim.deepcopy(_cached_adapter.headers)
-  headers["Authorization"] = "Bearer " .. _github_token.token
-
-  local ok, response = pcall(function()
-    return curl.get(url .. "/models", {
-      sync = true,
-      headers = headers,
-      insecure = config.adapters.opts.allow_insecure,
-      proxy = config.adapters.opts.proxy,
-    })
-  end)
-  if not ok then
-    log:error("Could not get the Copilot models from " .. url .. "/models.\nError: %s", response)
-    return {}
-  end
-
-  local ok, json = pcall(vim.json.decode, response.body)
-  if not ok then
-    log:error("Error parsing the response from " .. url .. "/models.\nError: %s", response.body)
-    return {}
-  end
-
-  local models = {}
-  for _, model in ipairs(json.data) do
-    if model.model_picker_enabled and model.capabilities.type == "chat" then
-      local choice_opts = {}
-
-      if model.capabilities.supports.streaming then
-        choice_opts.can_stream = true
-      end
-      if model.capabilities.supports.tool_calls then
-        choice_opts.can_use_tools = true
-      end
-      if model.capabilities.supports.vision then
-        choice_opts.has_vision = true
-      end
-
-      models[model.id] = { opts = choice_opts }
-    end
-  end
-
-  _cached_models = models
-  _cache_expires = utils.refresh_cache(_cache_file, config.adapters.opts.cache_models_for)
-
-  return models
-end
-
----Get Copilot usage statistics
----@return table|nil
-local function get_copilot_stats()
-  local dummy_adapter = { url = "" }
-  if not get_and_authorize_token(dummy_adapter) then
-    return nil
-  end
-
-  log:debug("Fetching Copilot usage statistics")
-
-  local ok, response = pcall(function()
-    return curl.get("https://api.github.com/copilot_internal/user", {
-      sync = true,
-      headers = {
-        Authorization = "Bearer " .. _oauth_token,
-        Accept = "*/*",
-        ["User-Agent"] = "CodeCompanion.nvim",
-      },
-      insecure = config.adapters.opts.allow_insecure,
-      proxy = config.adapters.opts.proxy,
-    })
-  end)
-  if not ok then
-    log:error("Could not get Copilot stats: %s", response)
-    return nil
-  end
-
-  local ok, json = pcall(vim.json.decode, response.body)
-  if not ok then
-    log:error("Error parsing Copilot stats response: %s", response.body)
-    return nil
-  end
-
-  return json
-end
-
----Show Copilot usage statistics in a floating window
----@return nil
-local function show_copilot_stats()
-  local stats = get_copilot_stats()
-  if not stats then
-    return vim.notify("Could not retrieve Copilot stats", vim.log.levels.ERROR)
-  end
-
-  local lines = {}
-  local ui = require("codecompanion.utils.ui")
-  table.insert(lines, "# 󰾞  GitHub Copilot Usage Statistics 󰾞 ")
-  table.insert(lines, "")
-
-  if stats.quota_snapshots.premium_interactions then
-    local premium = stats.quota_snapshots.premium_interactions
-    table.insert(lines, "##  Premium Interactions")
-    local used = premium.entitlement - premium.remaining
-    local usage_percent = premium.entitlement > 0 and (used / premium.entitlement * 100) or 0
-    table.insert(lines, string.format("   - Used: %d / %d (%.1f%%)", used, premium.entitlement, usage_percent))
-    table.insert(lines, string.format("   - Remaining: %d", premium.remaining))
-    table.insert(lines, string.format("   - Percentage: %.1f%%", premium.percent_remaining))
-    if premium.unlimited then
-      table.insert(lines, "   - Status: Unlimited ✨")
-    else
-      table.insert(lines, "   - Status: Limited")
-    end
-    table.insert(lines, "")
-  end
-
-  if stats.quota_snapshots.chat then
-    local chat = stats.quota_snapshots.chat
-    table.insert(lines, "## 󰭹 Chat")
-    if chat.unlimited then
-      table.insert(lines, "   - Status: Unlimited ✨")
-    else
-      local used = chat.entitlement - chat.remaining
-      local usage_percent = chat.entitlement > 0 and (used / chat.entitlement * 100) or 0
-      table.insert(lines, string.format("   - Used: %d / %d (%.1f%%)", used, chat.entitlement, usage_percent))
-    end
-    table.insert(lines, "")
-  end
-
-  if stats.quota_snapshots.completions then
-    local completions = stats.quota_snapshots.completions
-    table.insert(lines, "##  Completions")
-    if completions.unlimited then
-      table.insert(lines, "   - Status: Unlimited ✨")
-    else
-      local used = completions.entitlement - completions.remaining
-      local usage_percent = completions.entitlement > 0 and (used / completions.entitlement * 100) or 0
-      table.insert(lines, string.format("   - Used: %d / %d (%.1f%%)", used, completions.entitlement, usage_percent))
-    end
-  end
-  if stats.quota_reset_date then
-    table.insert(lines, "")
-    table.insert(lines, string.format("> Quota resets on: %s", stats.quota_reset_date))
-    table.insert(lines, "")
-  end
-
-  -- Create floating window
-  local float_opts = {
-    title = "󰍘 Copilot Stats",
-    lock = true,
-    relative = "editor",
-    row = "center",
-    col = "center",
-    window = {
-      width = 43,
-      height = math.min(#lines + 2, 20),
-    },
-    ignore_keymaps = false,
-  }
-  local _, winnr = ui.create_float(lines, float_opts)
-
-  local function get_usage_highlight(usage_percent)
-    if usage_percent >= 80 then
-      return "Error"
-    else
-      return "MoreMsg"
-    end
-  end
-  vim.api.nvim_win_call(winnr, function()
-    -- Usage percentages with color coding
-    local premium = stats.quota_snapshots.premium_interactions
-    if premium and not premium.unlimited then
-      local used = premium.entitlement - premium.remaining
-      local usage_percent = premium.entitlement > 0 and (used / premium.entitlement * 100) or 0
-      local highlight = get_usage_highlight(usage_percent)
-      vim.fn.matchadd(highlight, string.format("   - Used: %d / %d (%.1f%%)", used, premium.entitlement, usage_percent))
-    end
-  end)
-end
-
 ---@class Copilot.Adapter: CodeCompanion.Adapter
 return {
   name = "copilot",
@@ -386,10 +182,15 @@ return {
     ["Editor-Version"] = "Neovim/" .. vim.version().major .. "." .. vim.version().minor .. "." .. vim.version().patch,
   },
   get_copilot_stats = function()
-    return get_copilot_stats()
+    return copilot_helper.get_copilot_stats(get_and_authorize_token, _oauth_token)
   end,
   show_copilot_stats = function()
-    return show_copilot_stats()
+    -- we need to ensure initialize token if no chat request has been done
+    local dummy_adapter = { url = "" }
+    if not get_and_authorize_token(dummy_adapter) then
+      return nil
+    end
+    return copilot_helper.show_copilot_stats(get_and_authorize_token, _oauth_token)
   end,
   handlers = {
     ---Check for a token before starting the request
@@ -478,7 +279,7 @@ return {
       return openai.handlers.inline_output(self, data, context)
     end,
     on_exit = function(self, data)
-      reset()
+      copilot_helper.reset_cache()
       return openai.handlers.on_exit(self, data)
     end,
   },
@@ -492,7 +293,14 @@ return {
       ---@type string|fun(): string
       default = "gpt-4.1",
       choices = function(self)
-        return get_models(self)
+        -- Ensure token is available before getting models
+        if not _github_token then
+          local success = get_and_authorize_token(self)
+          if not success then
+            return { ["gpt-4.1"] = { opts = {} } } -- fallback
+          end
+        end
+        return copilot_helper.get_models(self, get_and_authorize_token, _github_token)
       end,
     },
     ---@type CodeCompanion.Schema
